+++ tests/Http/CurlHandlerTest.php
@@ -30,7 +30,7 @@ class CurlHandlerTest extends TestCase
     /** @var Client */
     private $client;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();
         $this->client = new Client(['apiKey' => 'QWERTY']);