+++ tests/Rest/ResourcesTest.php
@@ -34,7 +34,7 @@ class ResourcesTest extends TestCase
     /** @var Factory */
     private $factory;
 
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();