+++ tests/TestCase.php
@@ -60,7 +60,7 @@ abstract class TestCase extends Framework\TestCase
     /**
      * {@inheritdoc}
      */
-    protected function setUp()
+    protected function setUp(): void
     {
         parent::setUp();