+++ src/Client.php
@@ -22,7 +22,6 @@ use Psr\Http\Message\ResponseInterface as Response;
 use Psr\Log\LoggerInterface as Logger;
 use Rebilly\Http\CurlHandler;
 use Rebilly\Middleware\LogHandler;
-use Rebilly\Middleware\OrganizationIdHeader;
 use Rebilly\Rest\File;
 use RuntimeException;
 
@@ -218,12 +217,6 @@ final class Client
             $sessionToken = null;
         }
 
-        if (isset($organizationId)) {
-            $organizationId = new OrganizationIdHeader($organizationId);
-        } else {
-            $organizationId = null;
-        }
-
         if (isset($baseUrl)) {
             $baseUrl = ltrim($baseUrl, '/');
         } else {
@@ -253,6 +246,7 @@ final class Client
         } elseif (!is_callable($middleware)) {
             throw new RuntimeException('Middleware should be callable');
         }
+        $organizationId = isset($organizationId) ? $organizationId : null;
 
         $this->config = compact(
             'apiKey',
@@ -273,8 +267,7 @@ final class Client
 
         // Prepare middleware stack
         $this->middleware = new Middleware\CompositeMiddleware(
-            new Middleware\BaseUri($this->createUri($baseUrl)),
-            $organizationId,
+            new Middleware\BaseUri($this->createUri($baseUrl), $organizationId),
             new Middleware\UserAgent(self::SDK_VERSION),
             $authentication,
             $middleware,
@@ -503,6 +496,9 @@ final class Client
 
         // Remove version from URI
         $uri = preg_replace('#^/' . self::CURRENT_VERSION . '#', '', $uri);
+
+        // Remove organization id from URI
+        $uri = preg_replace('#^/organizations/[^/]+/(.+)$#', '$1', $uri);
 
         // Unserialize response body
         $content = json_decode((string) $response->getBody(), true) ?: [];