+++ src/Entities/Schema.php
@@ -132,6 +132,12 @@ final class Schema implements IteratorAggregate, ArrayAccess
             'subscriptions/{subscriptionId}/lead-source' => function (array $content) {
                 return new LeadSource($content);
             },
+            'subscription-pauses' => function (array $content) {
+                return new Collection(new SubscriptionPause(), $content);
+            },
+            'subscription-pauses/{subscriptionPauseId}' => function (array $content) {
+                return new SubscriptionPause($content);
+            },
             'payment-cards' => function (array $content) {
                 return new Collection(new PaymentCard(), $content);
             },