+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Security - in case of vulnerabilities.
 
 - [x] Added PHP 8.0 support.
 - [x] Added `SubscriptionService::delete()` method.
+- [x] Added `SubscriptionService::void()` method.
 
 ### Changed