+++ CHANGELOG.md
@@ -21,6 +21,7 @@ Security - in case of vulnerabilities.
 - [x] Added `SubscriptionService::delete()` method.
 - [x] Added `SubscriptionService::void()` method.
 - [x] Added `SubscriptionService::previewOrder()` method.
+- [x] Added `SubscriptionPauseService` service.
 
 ### Changed